fixed some minor bugs
