+++ react-ui/src/components/HOC/RequireAuth.js
