+++ react-ui/src/components/Landing.js
