+++ react-ui/src/components/NewUserContainer.js
